// src/app/core/models/catGeneral.model.ts
export interface catGeneral {
    uiRegistro: number;
    uiCategoria: number;
    sNombre: string;
    bActivo: boolean;
    bActivoFijo: boolean;
}

export class catGeneral {
    
}

// public int uiRegistro { get; set; }
// public int uiCategoria { get; set; }
// [Required]
// public string sNombre { get; set; }
// public bool bActivo { get; set; }

// public bool bActivoFijo { get; set; }